Interaction

ElectronGas.Interaction.KO — Method

function KO(q, n, param)

Dynamic part of KO interaction, with polarization approximated by zero temperature Π0. Returns the spin symmetric part and asymmetric part separately.

#Arguments:

q: momentum

n: matsubara frequency given in integer s.t. ωn=2πTn

param: other system parameters

ElectronGas.Interaction.RPA — Method

function RPA(q, n, param)

Dynamic part of RPA interaction, with polarization approximated by zero temperature Π0.

#Arguments:

q: momentum

n: matsubara frequency given in integer s.t. ωn=2πTn

param: other system parameters

ElectronGas.Interaction.coulomb — Method

function coulomb(q,param)

Bare interaction in momentum space. Coulomb interaction if Λs=0, Yukawa otherwise.

#Arguments:

q: momentum

param: other system parameters

ElectronGas.Interaction.landauParameterTakada — Method

TODO function landauParameterTakada(q, n, param)->exchange correlation kernel/Landau parameter

G factor with Takada's anzats. See Takada(doi:10.1103/PhysRevB.47.5202)(Eq.2.13-2.16). Now Landau parameter F. F(+-)=G(+-)*V

#Arguments:

q: momentum

n: matsubara frequency given in integer s.t. ωn=2πTn

param: other system parameters