Polarization

ElectronGas.Polarization.Polarization0_FiniteTemp — Method

function Polarization0_FiniteTemp(q, n, param, maxk=20, scaleN=20, minterval=1e-6, gaussN=10)

Finite temperature one-spin Π0 function for matsubara frequency and momentum. Analytically sum over transfer frequency and angular dependence of momentum, and numerically calculate integration of magnitude of momentum. Slower(~200μs) than Polarization0ZeroTemp. Assume G0^{-1} = iω_n - (k^2/(2m) - mu)

#Arguments:

q: momentum

n: matsubara frequency given in integer s.t. ωn=2πTn

param: other system parameters

maxk: optional, upper limit of integral -> maxk*kF

scaleN: optional, N of Log grid in LogDensedGrid, check CompositeGrids for more detail

minterval: optional, actual minterval of grid is this value times min(q,kF)

gaussN: optional, N of GaussLegendre grid in LogDensedGrid.

ElectronGas.Polarization.Polarization0_ZeroTemp — Method

function Polarization0_ZeroTemp(q, n, param)

Zero temperature one-spin Π0 function for matsubara frequency and momentum. For low temperature the finite temperature polarization could be approximated with this function to run faster(~200ns). Assume G0^{-1} = iωn - (k^2/(2m) - E_F).

#Arguments:

q: momentum

n: matsubara frequency given in integer s.t. ωn=2πTn

param: other system parameters

ElectronGas.Polarization.Polarization0wrapped — Method

function Polarization0wrapped(Euv, rtol, sgrid::SGT, param, pifunc = Polarization0_ZeroTemp) where{TGT, SGT}

Π0 function for matsubara frequency and momentum. Use Polarization0ZeroTemp by default, Polarization0FiniteTemp when pifunc is specified. Assume G0^{-1} = iωn - (k^2/(2m) - E_F). Return full polarization0 function stored in GreenFunc.GreenBasic.Green2DLR.

#Arguments:

Euv: Euv of DLRGrid

rtol: rtol of DLRGrid

sgrid: momentum grid

param: other system parameters

pifunc: single point Π0 function used. Require form with pifunc(k, n, param).